+++ src/Traits/CacheResults.php
@@ -7,44 +7,6 @@ namespace Weeks\Laravel\Repositories\Traits;
  */
 trait CacheResults
 {
-    // Implement Parent Methods
-
-    public function getAll($columns = null, $orderBy = 'created_at', $sort = 'DECS')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getPaginated($paged = 15, $orderBy = 'created_at', $sort = 'DECS')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getForSelect($data, $key = 'id', $orderBy = 'created_at', $sort = 'DECS')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getById($id)
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getItemByColumn($term, $column = 'slug')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getCollectionByColumn($term, $column = 'slug')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-    public function getActively($term, $column = 'slug')
-    {
-        return $this->processRequest(__FUNCTION__, func_get_args());
-    }
-
-
     /**
      * @return bool
      */
@@ -76,22 +38,21 @@ trait CacheResults
     /**
      * Perform the query and cache if required.
      *
+     * @param $callback
      * @param $method
      * @param $args
      * @return mixed
      */
-    protected function processRequest($method, $args)
+    protected function processCacheRequest($callback, $method, $args)
     {
         if ($this->isCaching() && ! in_array($method, $this->getIgnoredMethods())) {
 
             $key = $this->createCacheKey($method, $args);
 
-            return $this->getCache()->remember($key, $this->getCacheTtl(), function () use ($method, $args) {
-                return call_user_func_array(['parent', $method], $args);
-            });
+            return $this->getCache()->remember($key, $this->getCacheTtl(), $callback);
         }
 
-        return call_user_func_array(['parent', $method], $args);
+        return $callback();
     }
 
     /**