+++ tests/integration/ThrowsHttpExceptionsTraitTest.php
@@ -19,7 +19,6 @@ class ThrowsHttpExceptionsTraitTest extends Orchestra\Testbench\TestCase
      */
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('cache.default', 'array');
         $app['config']->set('database.default', 'testbench');
         $app['config']->set('database.connections.testbench', [
             'driver'   => 'sqlite',